+++ controllers/contactControllers.js
@@ -1,0 +1,83 @@
+const contacts = require("../models/contacts");
+const { addSchema } = require("../validation/contactValidationSchemas");
+const { HttpError } = require("../htttpError/httpError");
+
+const getAllContact = async (req, res, next) => {
+  try {
+    const result = await contacts.listContacts();
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const getOneContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contacts.getContactById(contactId);
+    if (!result) {
+      const error = new Error("Not found");
+      error.status = 404;
+      throw error;
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const newContact = async (req, res, next) => {
+  try {
+    const { error } = addSchema.validate(req.body, { abortEarly: false });
+    if (error) {
+      throw HttpError(404, error.message);
+    } else {
+      const { name, email, phone } = req.body;
+      const result = await contacts.addContact(name, email, phone);
+      res.status(201).json(result);
+    }
+  } catch (error) {
+    next(error);
+  }
+};
+
+const deleteContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contacts.removeContact(contactId);
+    if (!result) {
+      const error = new Error("Not found");
+      error.status = 404;
+      throw error;
+    }
+    res.status(200).json({ message: "contact deleted" });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const updateContact = async (req, res, next) => {
+  try {
+    const { error } = addSchema.validate(req.body, { abortEarly: false });
+    if (error) {
+      throw HttpError(404, error.message);
+    }
+    const { contactId } = req.params;
+    const result = await contacts.updateContact(contactId, req.body);
+    if (!result) {
+      const error = new Error("Not found");
+      error.status = 404;
+      throw error;
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+module.exports = {
+  getAllContact,
+  getOneContact,
+  newContact,
+  deleteContact,
+  updateContact,
+};
